+++ libsolidity/analysis/TypeChecker.cpp
@@ -86,6 +86,9 @@ bool TypeChecker::visit(ContractDefinition const& _contract)
 
 	ASTNode::listAccept(_contract.baseContracts(), *this);
 
+	if (StorageLayoutSpecifier const* layoutSpecifier = _contract.storageLayoutSpecifier())
+		layoutSpecifier->accept(*this);
+
 	for (auto const& n: _contract.subNodes())
 		n->accept(*this);