+++ test/libsolidity/syntaxTests/storageLayoutSpecifier/contract_named_at.sol
@@ -1,0 +1,4 @@
+contract at layout at 0x1234ABC { }
+// ====
+// stopAfter: parsing
+// ----