+++ test/libsolidity/syntaxTests/storageLayoutSpecifier/contract_named_layout.sol
@@ -1,0 +1,4 @@
+contract layout layout at 0x1234ABC { }
+// ====
+// stopAfter: parsing
+// ----