+++ test/libsolidity/syntaxTests/storageLayoutSpecifier/contract_with_members_named_layout_and_at.sol
@@ -1,0 +1,7 @@
+contract C layout at 0x1234 {
+    uint layout;
+    function at() public pure { }
+}
+// ====
+// stopAfter: parsing
+// ----