+++ test/libsolidity/syntaxTests/storageLayoutSpecifier/layout_specification_by_function.sol
@@ -1,0 +1,7 @@
+function f() pure returns (uint256) {
+    return 128;
+}
+contract C layout at f() { }
+// ====
+// stopAfter: parsing
+// ----