+++ test/libsolidity/syntaxTests/storageLayoutSpecifier/multi_token_expression.sol
@@ -1,0 +1,5 @@
+contract C layout at 5 minutes { }
+contract D layout at 2 gwei { }
+// ====
+// stopAfter: parsing
+// ----